[OT-347] [REFACTOR]: custom api 리팩토링
