+++ src/entities/custom/apis/getRadarApi.ts
@@ -1,0 +1,15 @@
+import { api } from "@/shared/api";
+import { ApiResponse } from "@/shared/types";
+
+export interface RadarResponse {
+  popularity: number;
+  immersion: number;
+  mania: number;
+  recency: number;
+  reWatch: number;
+}
+
+export const getRadarApi = async (): Promise<ApiResponse<RadarResponse>> => {
+  const response = await api.get<ApiResponse<RadarResponse>>("/radar");
+  return response.data;
+};
