+++ src/entities/custom/apis/putRadarApi.ts
@@ -1,0 +1,6 @@
+import { RadarResponse } from "@entities/custom/apis";
+import { api } from "@shared/api";
+
+export const putRadarApi = async (body: RadarResponse): Promise<void> => {
+  await api.put("/radar", body);
+};
