+++ src/entities/custom/hooks/useRadarRecommend.ts
@@ -1,9 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
-import { radarRecommendApi, GetRadarListParams } from "@entities/custom/apis";
+import {
+  GetRadarListParams,
+  getRadarRecommendApi,
+} from "@entities/custom/apis";
 
-export const useRadarRecommend = (params: GetRadarListParams) => { 
-    return useQuery({
-        queryKey: ["radar", "recommend"],
-        queryFn: () => radarRecommendApi(params),
-    })
-}
+export const useRadarRecommend = (params: GetRadarListParams) => {
+  return useQuery({
+    queryKey: ["radar", "recommend"],
+    queryFn: () => getRadarRecommendApi(params),
+  });
+};
